add,fix: ファイルを移動させディレクトリ整える，ファイル追加
